back/requirements.txt
flask
flask-cors
gunicorn
snowflake-connector-python
cryptography